requirements.txt
streamlit
st_pages
lxml
plotly.express